Add IDF parser module
