+++ converters/Makefile.mk
@@ -19,6 +19,7 @@ dist_pkgdata_DATA += converters/csv-table2glm-library.py
 dist_pkgdata_DATA += converters/csv-table2glm-object.py
 dist_pkgdata_DATA += converters/csv-visualcrossing-weather2glm-weather.py
 dist_pkgdata_DATA += converters/geometry.py
+dist_pkgdata_DATA += converters/idfparse.py
 dist_pkgdata_DATA += converters/csv2glm.py
 
 # gz -> glm
