+++ converters/autotest/test_idf2glm_output.glm
@@ -1,4 +1,4 @@
-// generated by /Users/dchassin/Documents/github/slacgismo/gridlabd/converters/idf2glm.py on 2022-11-17 14:58:05.533090
+// generated by /Users/dchassin/Documents/github/slacgismo/gridlabd/converters/idf2glm.py on 2022-11-17 15:36:26.654079
 // output to autotest/test_idf2glm_output.glm
 class building {
 	char32 weather;
